move BlockBuilder and Resolver impls for Miner into separate file
